+++ src/app.js
@@ -1,3 +1,4 @@
+
 import { waypoint } from 'waypoints'
 import { sticky } from 'sticky'